Factor out nested highlighting capability
